backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy>=3.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
sqlalchemy>=2.0.0
werkzeug==2.3.7
requests==2.31.0
google-generativeai>=0.5.0
protobuf>=4.25.3
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
